refactor: 헥사고날 아키텍처 기반 패키지 구조 재편 (#17)
- common/ → global/error/, global/exception/, global/handler/, global/filter/
- sample/ → sample/adapter/in/web/
- LoggingFilterConfig → FilterConfig 이름 변경
- 모든 import 경로 수정
- 기능 변경 없음 (순수 리팩토링)
