+++ src/main/java/com/example/exception/playground/global/exception/AccessDeniedException.java
@@ -1,4 +1,6 @@
-package com.example.exception.playground.common;
+package com.example.exception.playground.global.exception;
+
+import com.example.exception.playground.global.error.ErrorCode;
 
 public class AccessDeniedException extends BusinessException {
