+++ src/main/java/com/example/exception/playground/global/filter/FilterErrorResponseWriter.java
@@ -1,6 +1,10 @@
-package com.example.exception.playground.common;
+package com.example.exception.playground.global.filter;
 
+import com.example.exception.playground.global.error.ErrorCode;
+import com.example.exception.playground.global.error.ErrorResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.example.exception.playground.global.error.ErrorCode;
+import com.example.exception.playground.global.error.ErrorResponse;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.MDC;
