+++ src/main/java/com/example/exception/playground/global/filter/RateLimitFilter.java
@@ -1,8 +1,12 @@
-package com.example.exception.playground.common;
+package com.example.exception.playground.global.filter;
 
+import com.example.exception.playground.global.error.ErrorCode;
 import jakarta.servlet.FilterChain;
+import com.example.exception.playground.global.error.ErrorCode;
 import jakarta.servlet.ServletException;
+import com.example.exception.playground.global.error.ErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
+import com.example.exception.playground.global.error.ErrorCode;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.web.filter.OncePerRequestFilter;
