+++ src/main/java/com/example/exception/playground/global/handler/GlobalExceptionHandler.java
@@ -1,5 +1,8 @@
-package com.example.exception.playground.common;
+package com.example.exception.playground.global.handler;
 
+import com.example.exception.playground.global.error.ErrorCode;
+import com.example.exception.playground.global.error.ErrorResponse;
+import com.example.exception.playground.global.exception.BusinessException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
