+++ src/test/java/com/example/exception/playground/global/handler/GlobalExceptionHandlerIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.example.exception.playground.common;
+package com.example.exception.playground.global.handler;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
